requirements_gym.txt
absl-py==1.3.0
ale-py==0.7.5
alembic==1.10.4
asttokens==2.2.1
astunparse==1.6.3
autopep8==2.0.1
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.0.3
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.26.3
colorama==0.4.4
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
DataProperty==0.55.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
distro==1.7.0
elementpath==4.1.1
executing==1.2.0
Farama-Notifications==0.0.4
fasteners==0.14.1
filelock==3.12.0
flatbuffers==23.1.21
fonttools==4.39.3
fsspec==2023.4.0
future==0.18.2
gast==0.4.0
glfw==2.5.5
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
gym==0.21.0
gym-notices==0.0.8
h5py==3.8.0
httplib2==0.20.2
huggingface-hub==0.14.1
huggingface-sb3==2.2.4
idna==3.4
imageio==2.24.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.11.0
ipywidgets==8.0.4
jax-jumpy==1.0.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
lit==16.0.2
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mbstrdecoder==1.1.2
mdurl==0.1.2
meson==1.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
optuna==3.1.1
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
parso==0.8.3
pathvalidate==2.5.2
perlin-noise==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.0.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pyglet==2.0.4
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRep @ file:///homes/av1019/FYP/fyp/PyRep
pyRFC3339==1.1
pytablewriter==0.64.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
pyzmq==25.0.0
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
scikit-build==0.16.6
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
SecretStorage==3.3.1
SimpleITK==2.2.1
six==1.16.0
SQLAlchemy==2.0.10
stable-baselines3==1.8.0
stack-data==0.6.2
sympy==1.11.1
tabledata==1.3.1
tcolorpy==0.1.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
torch==2.0.0
torchmetrics==0.11.3
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typepy==1.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wadllib==1.3.6
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xdg==5
xmlschema==2.2.3
zipp==3.15.0
